builder,provision: move RegistryAuthConfig to dockercommon
