+++ provision/docker/provisioner_test.go
@@ -37,6 +37,7 @@ import (
 	"github.com/tsuru/tsuru/provision/docker/container"
 	internalNodeContainer "github.com/tsuru/tsuru/provision/docker/nodecontainer"
 	"github.com/tsuru/tsuru/provision/docker/types"
+	"github.com/tsuru/tsuru/provision/dockercommon"
 	"github.com/tsuru/tsuru/provision/nodecontainer"
 	"github.com/tsuru/tsuru/provision/provisiontest"
 	"github.com/tsuru/tsuru/quota"
@@ -287,7 +288,7 @@ func (s *S) TestDeploy(c *check.C) {
 		Repository: "tsuru/app-" + a.Name,
 		Tag:        "v1-builder",
 	}
-	err = s.p.Cluster().PullImage(pullOpts, mainDockerProvisioner.RegistryAuthConfig())
+	err = s.p.Cluster().PullImage(pullOpts, dockercommon.RegistryAuthConfig())
 	c.Assert(err, check.IsNil)
 	imgID, err := s.p.Deploy(&a, builderImgID, evt)
 	c.Assert(err, check.IsNil)
@@ -352,7 +353,7 @@ func (s *S) TestDeployWithLimiterActive(c *check.C) {
 		Repository: "tsuru/app-" + a.Name,
 		Tag:        "v1-builder",
 	}
-	err = s.p.Cluster().PullImage(pullOpts, mainDockerProvisioner.RegistryAuthConfig())
+	err = s.p.Cluster().PullImage(pullOpts, dockercommon.RegistryAuthConfig())
 	c.Assert(err, check.IsNil)
 	_, err = s.p.Deploy(&a, builderImgID, evt)
 	c.Assert(err, check.IsNil)
@@ -410,7 +411,7 @@ func (s *S) TestDeployWithLimiterGlobalActive(c *check.C) {
 		Repository: "tsuru/app-" + a.Name,
 		Tag:        "v1-builder",
 	}
-	err = s.p.Cluster().PullImage(pullOpts, mainDockerProvisioner.RegistryAuthConfig())
+	err = s.p.Cluster().PullImage(pullOpts, dockercommon.RegistryAuthConfig())
 	c.Assert(err, check.IsNil)
 	imgID, err := s.p.Deploy(&a, builderImgID, evt)
 	c.Assert(err, check.IsNil)
@@ -466,7 +467,7 @@ func (s *S) TestDeployQuotaExceeded(c *check.C) {
 		Repository: "tsuru/app-" + a.Name,
 		Tag:        "v1-builder",
 	}
-	err = s.p.Cluster().PullImage(pullOpts, mainDockerProvisioner.RegistryAuthConfig())
+	err = s.p.Cluster().PullImage(pullOpts, dockercommon.RegistryAuthConfig())
 	c.Assert(err, check.IsNil)
 	_, err = s.p.Deploy(&a, builderImgID, evt)
 	c.Assert(err, check.NotNil)
@@ -503,7 +504,7 @@ func (s *S) TestDeployCanceledEvent(c *check.C) {
 		Repository: "tsuru/app-" + app.GetName(),
 		Tag:        "v1-builder",
 	}
-	err = s.p.Cluster().PullImage(pullOpts, mainDockerProvisioner.RegistryAuthConfig())
+	err = s.p.Cluster().PullImage(pullOpts, dockercommon.RegistryAuthConfig())
 	c.Assert(err, check.IsNil)
 	done := make(chan bool)
 	go func() {
@@ -635,7 +636,7 @@ func (s *S) TestDeployErasesOldImages(c *check.C) {
 		Repository: "tsuru/app-" + a.Name,
 		Tag:        "v1-builder",
 	}
-	err = s.p.Cluster().PullImage(pullOpts, mainDockerProvisioner.RegistryAuthConfig())
+	err = s.p.Cluster().PullImage(pullOpts, dockercommon.RegistryAuthConfig())
 	c.Assert(err, check.IsNil)
 	_, err = s.p.Deploy(&a, builderImgID, evt)
 	c.Assert(err, check.IsNil)
@@ -655,7 +656,7 @@ func (s *S) TestDeployErasesOldImages(c *check.C) {
 		Repository: "tsuru/app-" + a.Name,
 		Tag:        "v2-builder",
 	}
-	err = s.p.Cluster().PullImage(pullOpts, mainDockerProvisioner.RegistryAuthConfig())
+	err = s.p.Cluster().PullImage(pullOpts, dockercommon.RegistryAuthConfig())
 	c.Assert(err, check.IsNil)
 	_, err = s.p.Deploy(&a, builderImgID, evt)
 	c.Assert(err, check.IsNil)
@@ -745,7 +746,7 @@ func (s *S) TestDeployErasesOldImagesWithLongHistory(c *check.C) {
 		Repository: "tsuru/app-" + a.Name,
 		Tag:        "v1-builder",
 	}
-	err = s.p.Cluster().PullImage(pullOpts, mainDockerProvisioner.RegistryAuthConfig())
+	err = s.p.Cluster().PullImage(pullOpts, dockercommon.RegistryAuthConfig())
 	c.Assert(err, check.IsNil)
 	_, err = s.p.Deploy(&a, builderImgID, evt)
 	c.Assert(err, check.IsNil)
@@ -765,7 +766,7 @@ func (s *S) TestDeployErasesOldImagesWithLongHistory(c *check.C) {
 		Repository: "tsuru/app-" + a.Name,
 		Tag:        "v2-builder",
 	}
-	err = s.p.Cluster().PullImage(pullOpts, mainDockerProvisioner.RegistryAuthConfig())
+	err = s.p.Cluster().PullImage(pullOpts, dockercommon.RegistryAuthConfig())
 	c.Assert(err, check.IsNil)
 	_, err = s.p.Deploy(&a, builderImgID, evt)
 	c.Assert(err, check.IsNil)
@@ -787,7 +788,7 @@ func (s *S) TestDeployErasesOldImagesWithLongHistory(c *check.C) {
 		Repository: "tsuru/app-" + a.Name,
 		Tag:        "v3-builder",
 	}
-	err = s.p.Cluster().PullImage(pullOpts, mainDockerProvisioner.RegistryAuthConfig())
+	err = s.p.Cluster().PullImage(pullOpts, dockercommon.RegistryAuthConfig())
 	c.Assert(err, check.IsNil)
 	_, err = s.p.Deploy(&a, builderImgID, evt)
 	c.Assert(err, check.IsNil)
@@ -891,7 +892,7 @@ func (s *S) TestDeployImageID(c *check.C) {
 		Repository: "tsuru/app-" + a.Name,
 		Tag:        "v1",
 	}
-	err = s.p.Cluster().PullImage(pullOpts, mainDockerProvisioner.RegistryAuthConfig())
+	err = s.p.Cluster().PullImage(pullOpts, dockercommon.RegistryAuthConfig())
 	c.Assert(err, check.IsNil)
 	_, err = s.p.Deploy(&a, builderImgID, evt)
 	c.Assert(err, check.IsNil)
@@ -975,7 +976,7 @@ func (s *S) TestProvisionerDestroyRemovesImage(c *check.C) {
 		Repository: "tsuru/app-" + a.Name,
 		Tag:        "v1-builder",
 	}
-	err = s.p.Cluster().PullImage(pullOpts, mainDockerProvisioner.RegistryAuthConfig())
+	err = s.p.Cluster().PullImage(pullOpts, dockercommon.RegistryAuthConfig())
 	c.Assert(err, check.IsNil)
 	_, err = s.p.Deploy(&a, builderImgID, evt)
 	c.Assert(err, check.IsNil)
